+++ changelog.md
@@ -1,6 +1,7 @@
 ### ?, ? - V1.4.1.
 
 - Increase undo history size.
+- Organize presets in folders.
 
 ### March 18, 2024 - V1.4.0.